chore: move the gateway audio-stream proxy to its own PR
The gateway half of this work grew into a change to a shipped, security
sensitive proxy: a shared authorization and frame pump under /v1/stt/stream and
/v1/watch/stream, a guardian pin, and a port of live-voice-websocket.ts onto
the shared helper. None of that is what this PR's title leads a reviewer to
expect, and it is reviewed better on its own.

The cut is by directory and the two sides compile independently: the client
dials /v1/watch/stream by URL rather than importing anything from gateway/.
The ARCHITECTURE.md section travels with the gateway PR because the auth
posture it documents is that PR's subject.
