+++ gateway/src/index.ts
@@ -50,11 +50,6 @@ import {
   type SttStreamSocketData,
 } from "./http/routes/stt-stream-websocket.js";
 import {
-  createWatchStreamWebsocketHandler,
-  getWatchStreamWebsocketHandlers,
-  type WatchStreamSocketData,
-} from "./http/routes/watch-stream-websocket.js";
-import {
   createSpeechRelayUpgradeHandler,
   getSpeechRelayWebsocketHandlers,
   type SpeechRelaySocketData,
@@ -311,14 +306,6 @@ function isSttStreamSocketData(data: unknown): data is SttStreamSocketData {
   );
 }
 
-function isWatchStreamSocketData(data: unknown): data is WatchStreamSocketData {
-  return (
-    !!data &&
-    typeof data === "object" &&
-    (data as { wsType?: unknown }).wsType === "watch-stream"
-  );
-}
-
 function isLiveVoiceSocketData(data: unknown): data is LiveVoiceSocketData {
   return (
     !!data &&
@@ -548,7 +535,6 @@ async function main() {
     credentials: credentialCache,
   });
   const handleSttStreamWs = createSttStreamWebsocketHandler(config);
-  const handleWatchStreamWs = createWatchStreamWebsocketHandler(config);
   const handleLiveVoiceWs = createLiveVoiceWebsocketHandler(config);
   const handleSpeechRelaySttWs = createSpeechRelayUpgradeHandler(
     config,
@@ -563,7 +549,6 @@ async function main() {
   const twilioMediaStreamWebsocketHandlers = getMediaStreamWebsocketHandlers();
   const pluginWebhookWebsocketHandlers = getPluginWebhookWebsocketHandlers();
   const sttStreamWebsocketHandlers = getSttStreamWebsocketHandlers();
-  const watchStreamWebsocketHandlers = getWatchStreamWebsocketHandlers();
   const liveVoiceWebsocketHandlers = getLiveVoiceWebsocketHandlers();
   const speechRelayWebsocketHandlers = getSpeechRelayWebsocketHandlers();
   const { handler: handleWhatsAppWebhook, dedupCache: whatsappDedupCache } =
@@ -1872,10 +1857,6 @@ async function main() {
           sttStreamWebsocketHandlers.open(ws as never);
           return;
         }
-        if (isWatchStreamSocketData(ws.data)) {
-          watchStreamWebsocketHandlers.open(ws as never);
-          return;
-        }
         if (isLiveVoiceSocketData(ws.data)) {
           liveVoiceWebsocketHandlers.open(ws as never);
           return;
@@ -1899,10 +1880,6 @@ async function main() {
           sttStreamWebsocketHandlers.message(ws as never, message);
           return;
         }
-        if (isWatchStreamSocketData(ws.data)) {
-          watchStreamWebsocketHandlers.message(ws as never, message);
-          return;
-        }
         if (isLiveVoiceSocketData(ws.data)) {
           liveVoiceWebsocketHandlers.message(ws as never, message);
           return;
@@ -1924,10 +1901,6 @@ async function main() {
         }
         if (isSttStreamSocketData(ws.data)) {
           sttStreamWebsocketHandlers.close(ws as never, code, reason);
-          return;
-        }
-        if (isWatchStreamSocketData(ws.data)) {
-          watchStreamWebsocketHandlers.close(ws as never, code, reason);
           return;
         }
         if (isLiveVoiceSocketData(ws.data)) {
@@ -2148,12 +2121,6 @@ async function main() {
 
     if (url.pathname === "/v1/stt/stream") {
       const upgradeResult = handleSttStreamWs(req, server);
-      if (upgradeResult !== undefined) return upgradeResult;
-      return undefined as unknown as Response;
-    }
-
-    if (url.pathname === "/v1/watch/stream") {
-      const upgradeResult = await handleWatchStreamWs(req, server);
       if (upgradeResult !== undefined) return upgradeResult;
       return undefined as unknown as Response;
     }
